Comply with the CC BY-NC-ND 4.0 license of LAFAN
